chore(@clayui/management-toolbar): update ManagementToolbar component stories
